4.2.6. блокирует фильтр WIP
