+++ src/presenter/filter-presenter.js
@@ -18,7 +18,13 @@ export default class FilterPresenter {
   init() {
     const filters = this._getFilters();
     const prevFilterView = this._filtersView;
-    this._filtersView = new FilterView(filters, this._filterModel.getActiveFilter());
+
+    const filterType = this._filterModel.getActiveFilter();
+    const length = this._pointsModel.getPoints(filterType).length;
+console.log('0o0', length)
+
+
+    this._filtersView = new FilterView(filters, filterType, length);
     this._filtersView.setFilterChangeHandler(this._handleFilterTypeChange);
     if (prevFilterView === null) {
       render(this._filterContainer, this._filtersView, RenderPosition.BEFOREEND);
